版本升级到1.6.1 日志工具类支持打印日志到自定义文件 或者自带的沙盒路径Cache下 可兼容10.0

+++ easyComLibrary/src/main/java/com/zs/easy/common/utils/LogUtil.java
@@ -3,6 +3,7 @@ package com.zs.easy.common.utils;
 import android.util.Log;
 
 import com.zs.easy.common.constants.EasyConstants;
+import com.zs.easy.common.constants.EasyVariable;
 
 import java.io.BufferedWriter;
 import java.io.File;
@@ -197,7 +198,7 @@ public class LogUtil {
     }
 
     /**
-     * 打开日志文件并写入日志 自定义文件名
+     * 打开日志文件并写入日志 自定义文件名 默认在sd卡根目录
      *
      * @return
      **/
@@ -207,7 +208,6 @@ public class LogUtil {
         }
         Log.i(tag, text);
         Date nowtime = new Date();
-        String needWriteFiel = logfile.format(nowtime);
         String needWriteMessage = myLogSdf.format(nowtime) + "    " + tag + "    " + text;
 
         // 取得日志存放目录
@@ -216,6 +216,69 @@ public class LogUtil {
             try {
                 // 打开文件
                 File file = new File(path + File.separator + fileName);
+                FileWriter filerWriter = new FileWriter(file, true);// 后面这个参数代表是不是要接上文件中原来的数据，不进行覆盖
+                BufferedWriter bufWriter = new BufferedWriter(filerWriter);
+                bufWriter.write(needWriteMessage);
+                bufWriter.newLine();
+                bufWriter.close();
+                filerWriter.close();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+    }
+
+    /**
+     * 打开日志文件并写入日志 自定义文件名 存在Android data 包名 Cache下面
+     * 如果sd卡不可用就存在 data  data  包名  Cache下面
+     *
+     * @return
+     **/
+    public static void writeLogtoCache(String tag, String text, String fileName) {// 新建或打印到日志文件
+        if (!logable) {
+            return;
+        }
+        Log.i(tag, text);
+        Date nowtime = new Date();
+        String needWriteMessage = myLogSdf.format(nowtime) + "    " + tag + "    " + text;
+
+        // 取得日志存放目录
+        String path = DeviceUtil.getCachePath(EasyVariable.mContext);
+        if (path != null && !"".equals(path)) {
+            try {
+                // 打开文件
+                File file = new File(path + File.separator + fileName);
+                if (!file.exists()) {
+                    file.mkdir();
+                }
+                FileWriter filerWriter = new FileWriter(file, true);// 后面这个参数代表是不是要接上文件中原来的数据，不进行覆盖
+                BufferedWriter bufWriter = new BufferedWriter(filerWriter);
+                bufWriter.write(needWriteMessage);
+                bufWriter.newLine();
+                bufWriter.close();
+                filerWriter.close();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+    }
+
+    /**
+     * 打开日志文件并写入日志 自定义文件
+     *
+     * @return
+     **/
+    public static void writeLogtoFile(String tag, String text, File file) {// 新建或打印到日志文件
+        if (!logable) {
+            return;
+        }
+        Log.i(tag, text);
+        Date nowtime = new Date();
+        String needWriteMessage = myLogSdf.format(nowtime) + "    " + tag + "    " + text;
+
+        if (file != null && file.exists()) {
+            try {
+                // 打开文件
                 FileWriter filerWriter = new FileWriter(file, true);// 后面这个参数代表是不是要接上文件中原来的数据，不进行覆盖
                 BufferedWriter bufWriter = new BufferedWriter(filerWriter);
                 bufWriter.write(needWriteMessage);
